requirements.txt
Django==5.2.4
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
django-cors-headers>=3.10.0
numpy>=1.21.0
matplotlib>=3.4.0